rag-llm-aws/requirements.txt
flask
langchain
langchain-openai
langchain-groq
langchain-google-genai
python-dotenv
pydantic
pymongo
pandas
numpy
joblib
flask-login
sagemaker
boto3